# scx_p2dq: Refactor dispatch under saturation